Fix minor compiler warning/error.
[[originally from svn r17203]]

+++ utility/support.c
@@ -466,7 +466,7 @@ char *fc_strrep_resize(char *str, size_t *len, const char *search,
   RETURN_VAL_IF_FAIL_MSG(fc_strrep(str, (*len), search, replace), NULL,
                          "Can't replace '%s' by '%s' in '%s'. To small "
                          "size after reallocation: %lu.", search, replace,
-                         str, *len);
+                         str, (long unsigned int)*len);
 
   return str;
 }
